src/tools/ise-migration/src/requirements.txt
aiosignal==1.2.0
alchemy-mock==0.4.3
altgraph==0.17.2
aniso8601==9.0.1
ansicon==1.89.0
anyio==3.6.2
ariadne==0.16.1
astroid==2.9.3
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
azure-common==1.1.27
azure-core==1.26.4
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-storage-file-share==12.12.0
bcrypt==4.0.1
bidict==0.22.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.4
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docopt==0.6.2
ezenv==0.92
Faker==15.3.3
fastapi==0.85.1
filelock==3.6.0
Flask==2.2.2
fonttools==4.38.0
fpdf==1.7.2
fpdf2==2.6.0
frozenlist==1.3.1
future==0.18.2
gevent==22.10.1
gevent-websocket==0.10.1
ggshield==1.10.8
git-filter-repo==2.38.0
graphene==3.1.1
graphene-pydantic==0.4.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==1.1.3.post0
h11==0.14.0
httplib2==0.20.4
httptools==0.5.0
hypothesis==6.36.2
identify==2.4.11
idna==3.3
iniconfig==1.1.1
install==1.3.5
isodate==0.6.1
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
jinxed==1.2.0
jsonpickle==2.2.0
lazy-object-proxy==1.7.1
loguru==0.6.0
MarkupSafe==2.1.1
marshmallow==3.14.1
mccabe==0.6.1
mock==4.0.3
mock-alchemy==0.2.5
msal==1.16.0
msal-extensions==0.3.1
msrest==0.6.21
multidict==6.0.2
nest-asyncio==1.5.6
nodeenv==1.6.0
oauthlib==3.1.1
openapi3==1.6.5
packaging==21.3
pandas==1.4.4
paramiko==2.12.0
pefile==2022.5.30
Pillow==9.3.0
platformdirs==2.5.0
pluggy==1.0.0
portalocker==2.3.2
pre-commit==2.17.0
promise==2.3
psycopg2==2.9.3
py==1.11.0
py-automapper==1.2.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.2
pygitguardian==1.3.1
Pygments==2.12.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
PyJWT==2.3.0
pylint==2.12.2
PyNaCl==1.5.0
pyparsing==3.0.7
Pypubsub==4.0.3
pysftp==0.2.9
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
redis==4.4.0
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
rich==12.4.4
rq==1.11.1
Rx==1.6.1
shiv==1.0.1
shortuuid==1.0.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
starlette==0.20.4
starlette-graphene3==0.6.0
toml==0.10.2
tomli==2.0.1
typer==0.4.1
typing_extensions==4.5.0
uamqp==1.5.1
urllib3==1.26.13
uvicorn==0.19.0
validators==0.20.0
virtualenv==20.18.0
virtualenvwrapper-win==1.2.7
watchfiles==0.18.0
websockets==10.3
Werkzeug==2.2.2
win32-setctime==1.1.0
wrapt==1.13.3
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.5.0
